requirements.txt
streamlit
python-docx
requests
Pillow